requirements.txt
biopython==1.73
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.0
Python==3.7.4
scikit-learn==0.22.1
seaborn==0.10.0
tensorflow==2.1.0